backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
djoser==2.1.0
psycopg2-binary==2.9.3
Pillow==9.0.0
django-redis==5.4.0
PyYAML==6.0
django-filter==23.1
python-dotenv==0.20.0